audioPrediction/requirements.txt
# Use tensorflow-cpu for better Lambda compatibility and smaller size
tensorflow-cpu==2.14.0

# BirdNet Library and its core dependencies
birdnetlib==0.11.0
librosa==0.10.1
pydub==0.25.1
numpy==1.24.3
soundfile==0.12.1
resampy

# Additional dependencies that might be needed
boto3==1.34.0
botocore==1.34.0